Use . instead of source for /bin/sh compatibility, extract PythonCommandHelper
