+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/python/PythonBulkImportDriver.java
@@ -16,9 +16,6 @@
 
 package sleeper.systemtest.drivers.python;
 
-import sleeper.clients.util.ClientUtils;
-import sleeper.clients.util.CommandPipeline;
-import sleeper.clients.util.CommandPipelineRunner;
 import sleeper.systemtest.drivers.instance.SleeperInstanceContext;
 
 import java.io.IOException;
@@ -32,16 +29,17 @@ import static sleeper.configuration.properties.instance.IngestProperty.INGEST_SO
 
 public class PythonBulkImportDriver {
     private final SleeperInstanceContext instance;
-    private final CommandPipelineRunner pipelineRunner = ClientUtils::runCommandInheritIO;
+    private final PythonCommandHelper pythonCommandHelper;
     private final Path pythonDir;
 
     public PythonBulkImportDriver(SleeperInstanceContext instance, Path pythonDir) {
         this.instance = instance;
+        this.pythonCommandHelper = new PythonCommandHelper(pythonDir);
         this.pythonDir = pythonDir;
     }
 
     public void emrServerless(String jobId, String... files) throws IOException, InterruptedException {
-        runInVenv(pipeline(command(Stream.concat(
+        pythonCommandHelper.runInVenv(pipeline(command(Stream.concat(
                         Stream.of("python3",
                                 pythonDir.resolve("test/bulk_import_files_from_s3.py").toString(),
                                 "--instance", instance.getInstanceProperties().get(ID),
@@ -52,19 +50,5 @@ public class PythonBulkImportDriver {
                         Stream.of(files)
                                 .map(file -> instance.getInstanceProperties().get(INGEST_SOURCE_BUCKET) + "/" + file))
                 .toArray(String[]::new))));
-    }
-
-    public void runInVenv(CommandPipeline pipeline) throws IOException, InterruptedException {
-        pipelineRunner.run(activateVenv(pythonDir));
-        pipelineRunner.run(pipeline);
-        pipelineRunner.run(deactivateVenv());
-    }
-
-    public static CommandPipeline activateVenv(Path pythonDir) {
-        return pipeline(command("source", pythonDir.resolve("env/bin/activate").toString()));
-    }
-
-    public static CommandPipeline deactivateVenv() {
-        return pipeline(command("deactivate"));
     }
 }
